Ungly handing of top level library
I've made the ability to set the top level library to get around issues
where the use of the library was really unitnended. Like 99% of the
files everything references 'work' but then one references
'xil_defaultlib' instead.

I'm gong to have a think about how better the handle this issue. Using
the command line option in this file is a temprary work around