Fixed the way components are handled
This fixers the way components are handle. If a commponent cannot be
found now only a warning is produced. There have also been multiple new
tags added

** ignore_components **
This will stop a warning from coming up with the component cannot be
found.

** package_file_skip_order **
This will add the package file so that the component declartaion can be
used from it but will not add the packege to the compile order. This is
useful for Xilinx UNSIM and XPM VCOMP libraries. Where it will not
compliate that the componet cannot be found (becuase it can be found in
the added library) but it will not add the library to the compile order
at the end. (as these libraries do not normally need to compile but are
precompiled)

I've also modifed to that it can take a .json config file. As well as
the current .toml file.

So the ** sub_toml ** tag has now been moved to just ** sub **