Starting option for tracking file version numbers
I've added an option for keeping track of versions which will be passed
as the file type.

Inital use case if for identifing VHDL2008 from VHDL1993. This will
probably need to be refined later. Also documentation needs to be added.

Finally I've added a default VHDL library ignore for default libraries
ieee and std.