+X_XCI and +filtering conflicts on Vivado version
I've added a new file type Xilinx XCI. For the FIR filter IP I've
also made it add direct dependencies below it used for holding
coefficient. This will need to be expanded for other IP in the future.

I've now have the option of fitler out conflicts with X_BD and X_XCI
based on the version of Vivado and device you are targetting